requirements.txt
requests
python-dotenv
mysqlclient
sqlalchemy
mysql-connector-python